# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Changeorg Laravel API Rest Documentation",
        "description": "Changeorg Laravel API Rest Documentation",
        "contact": {
            "email": "ilarranaga@jesuitasformacion.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://laravel-changeorg-api.herokuapp.com/",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Sign in",
                "description": "Login by email, password",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Good credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/peticiones": {
            "get": {
                "tags": [
                    "Peticiones"
                ],
                "summary": "Get list of peticiones",
                "description": "Returns list of petiiones",
                "operationId": "b91666912c965c676b40d2d7d767954b",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#Peticiones"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/mispeticiones/": {
            "get": {
                "tags": [
                    "Peticiones"
                ],
                "summary": "Get list of peticiones of the logged in user",
                "description": "Returns list of peticiones of the logged in user",
                "operationId": "3871972bac61a1afd66ad9ca7b638fad",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#Peticiones"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/peticiones/{id}": {
            "get": {
                "summary": "Muestra el detalle de una petición",
                "operationId": "2c711e8f3241fad7fb3766afc631b04e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Peticione id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar la petición."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "put": {
                "tags": [
                    "Peticiones"
                ],
                "summary": "Update existing peticione",
                "description": "Returns updated peticione data",
                "operationId": "f65dfdf1140faa8f9eaf32275765e293",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Peticione id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#Peticione"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#Peticione"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/users/firmas": {
            "get": {
                "summary": "Mostrar peticiones firmadas por el usuario logueado",
                "operationId": "80dbcadf913999ea1ada57c555d17f36",
                "responses": {
                    "200": {
                        "description": "Mostrar todas las peticiones."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        }
    },
    "components": {
        "responses": {
            "401": {
                "description": "Wrong credentials response",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "Unauthorized. Either email or password is wrong."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "ChangeOrg API",
            "description": "API Endpoints"
        }
    ]
}